requirements.txt
pandas
pyarrow
pydantic
pyyaml
rich
